Fix check for auth/sso to happen only on success
The waa2wls protocol document doesn't make it entirely clear, but the
condition for sso, "REQUIRED if 'auth' is empty", only makes sense if
authentication was successful -- if authentication didn't happen, then
of course we cannot provide information on which auth method(s) were
used, now or in the past.

The protocol document probably needs amending to correct this.
Currently, it reads as if sso MUST be provided if auth is empty (which
itself MUST be empty if authentication was not successful).